Completing migration of feature based visualizations
