+++ pyreal/visualize/__init__.py
@@ -1,4 +1,3 @@
-from pyreal.visualize.feature_based_viz import plot_top_contributors
+from pyreal.visualize.feature_based_viz import plot_top_contributors, swarm_plot
 
-__all__ = [
-    "plot_top_contributors"]
+__all__ = ["plot_top_contributors", "swarm_plot"]
